change user routes back to auth
